+++ src/main/java/com/example/pigonair/global/config/security/UserDetailsImpl.java
@@ -4,10 +4,9 @@ import java.util.ArrayList;
 import java.util.Collection;
 
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import com.example.pigonair.member.entity.Member;
+import com.example.pigonair.domain.member.entity.Member;
 
 
 public class UserDetailsImpl implements UserDetails {
